Prepare splitting Console onto files
