+++ lib/console/loader.rb
@@ -1,0 +1,5 @@
+class Console
+  module Loader
+    # foobar
+  end
+end
